requirements.txt
aniso8601==6.0.0
astroid==1.6.6
backports.functools-lru-cache==1.5
beautifulsoup4==4.7.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
configparser==3.7.4
Django==1.11.22
enum34==1.1.6
Flask==1.1.1
futures==3.2.0
graphene==2.1.7
graphql-core==2.2.1
graphql-relay==2.0.0
gunicorn==19.9.0
idna==2.8
isort==4.3.20
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.16.4
pipenv==2018.11.26
promise==2.2.1
pylint==1.9.4
python-dotenv==0.10.3
pytz==2019.1
requests==2.22.0
Rx==1.6.1
scikit-learn==0.20.3
scipy==1.2.2
selenium==3.141.0
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
soupsieve==1.9.2
typing==3.7.4
urllib3==1.25.3
virtualenv==16.6.1
virtualenv-clone==0.5.3
Werkzeug==0.15.5
wrapt==1.11.2
